simplify integration of s3 upload payload signing
