+++ Sources/SmileID/Classes/DocumentVerification/Model/OrchestratedDocumentVerificationViewModel.swift
@@ -1,6 +1,5 @@
 import Combine
 import SwiftUI
-import SmileIDSecurity
 
 enum DocumentCaptureFlow: Equatable {
     case frontDocumentCapture
@@ -180,7 +179,9 @@ class IOrchestratedDocumentVerificationViewModel<T, U: JobResult>: ObservableObj
                     livenessImages: livenessFiles
                 )
                 allFiles.append(info)
-                allFiles = try addSecurityInfo(files: allFiles)
+                if let securityInfoJson = try LocalStorage.addSecurityInfo(jobId: jobId, files: allFiles) {
+                    allFiles.append(contentsOf: [securityInfoJson])
+                }
                 let zipData = try LocalStorage.zipFiles(at: allFiles)
                 self.savedFiles = DocumentCaptureResultStore(
                     allFiles: allFiles,
@@ -314,39 +315,6 @@ class IOrchestratedDocumentVerificationViewModel<T, U: JobResult>: ObservableObj
                 self.onError(error: error)
             }
         }
-    }
-
-    private func addSecurityInfo(files: [URL]) throws -> [URL] {
-        // Sort the files by their last path component (filename) alphabetically
-        let sortedFiles = files.sorted { $0.lastPathComponent < $1.lastPathComponent }
-
-        var data = Data()
-        for file in sortedFiles {
-            var fileData = try Data(contentsOf: file)
-            if file.lastPathComponent.contains(FileType.selfie.rawValue) ||
-                file.lastPathComponent.contains(FileType.liveness.rawValue) ||
-                file.lastPathComponent.contains(FileType.documentFront.rawValue) ||
-                file.lastPathComponent.contains(FileType.documentBack.rawValue) {
-                fileData = fileData.base64EncodedData()
-            }
-            data.append(fileData)
-        }
-
-        let timestamp = Date().toISO8601WithMilliseconds()
-        let mac = try SmileIDCryptoManager.shared.sign(
-            timestamp: timestamp,
-            payload: data
-        )
-
-        let securityInfo = SecurityInfo(
-            timestamp: timestamp,
-            mac: mac
-        )
-        let securityInfoJson = try LocalStorage.createSecurityInfoFile(
-            jobId: jobId,
-            securityInfo: securityInfo
-        )
-        return files + [securityInfoJson]
     }
 
     /// If stepToRetry is ProcessingScreen, we're retrying a network issue, so we need to kick off
